fix use of "tp ~ ~ ~" for tp nerfing
That is, "tp user X Y Z" used to be relative to the user location,
so you could tp user ~ ~ ~ to teleport a user to their current location.

Now it is the same as "teleport" and relative to executor location,
which means for the console it's spawn point.  Ugh.

Fortunately, `execute` lets us run it as the user...

...but the position is still spawn unless we run it *at* the user, too.